# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /moderation/oauth/discord:
    get:
      operationId: moderationDiscordAuthorize
      tags: [moderation]
      responses:
        "302": { description: Redirects to Discord with browser-bound OAuth state }
        "503": { description: Discord OAuth is not configured }
...
  /moderation/storage/uploads/request-url:
    post:
      operationId: requestModerationStorageUploadUrl
      tags: [moderation]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ModerationStorageUploadRequest" } } }
      responses:
        "200":
          description: Owner-authorized temporary upload URL and object path
          content: { application/json: { schema: { $ref: "#/components/schemas/ModerationStorageUploadResponse" } } }
        "400": { description: Invalid image metadata }
        "401": { description: Moderation session is required }
        "403": { description: Guild ownership is required }
...
components:
  schemas:
...
    ModerationStorageUploadRequest:
      type: object
      additionalProperties: false
      required: [guildId, name, size, contentType]
      properties:
        guildId: { type: string, minLength: 5, maxLength: 25 }
        name: { type: string, minLength: 1, maxLength: 200 }
        size: { type: integer, minimum: 1, maximum: 10485760 }
        contentType: { type: string, pattern: "^image/(jpeg|png|gif|webp)$" }
    ModerationStorageUploadResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: "#/components/schemas/ModerationStorageUploadRequest" }
...